requirements.txt
absl-py==0.6.1
astor==0.7.1
captcha==0.3
certifi==2018.10.15
chardet==3.0.4
claptcha==0.3.3
cycler==0.10.0
cython==0.29.3
easy-captcha==1.3
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
numpy==1.14.5
object-detection==0.1
pandas==0.23.4
Pillow==5.3.0
protobuf==3.6.1
pygame==1.9.4
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.1
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Shapely==1.6.4.post1
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wincertstore==0.2
